test: factor out shared code from sns tests (#803)
