+++ __tests__/api-sns.test.ts
@@ -3,26 +3,33 @@ jest.mock("../app/utils/http", () => {
   return { httpGet: jest.fn() };
 });
 
+import {
+  createS3Event,
+  createSNSMessage,
+  createTestAtlasData,
+  setUpAwsConfig,
+  SNS_MESSAGE_DEFAULTS,
+  SQL_QUERIES,
+  TEST_AWS_CONFIG,
+  TEST_ERROR_MESSAGE_INVALID_SIGNATURE,
+  TEST_FILE_PATHS,
+  TEST_MODULE_SNS_VALIDATOR,
+  TEST_S3_BUCKET,
+  TEST_SIGNATURE,
+  TEST_SIGNATURE_INVALID,
+  TEST_SIGNATURE_VALID,
+  TEST_TIMESTAMP,
+  TEST_TIMESTAMP_ALT,
+  TEST_VERSION_IDS,
+  validateTestSnsMessage,
+} from "../testing/sns-testing";
+
 // Set up AWS resource configuration BEFORE any other imports
-const TEST_S3_BUCKET = "hca-atlas-tracker-data-dev";
-const TEST_GUT_ATLAS_ID = "550e8400-e29b-41d4-a716-446655440000";
-const TEST_S3_EVENT_NAME = "s3:ObjectCreated:Put";
-const TEST_AWS_CONFIG = {
-  s3_buckets: [TEST_S3_BUCKET],
-  sns_topics: [
-    "arn:aws:sns:us-east-1:123456789012:hca-atlas-tracker-s3-notifications",
-  ],
-};
-process.env.AWS_RESOURCE_CONFIG = JSON.stringify(TEST_AWS_CONFIG);
+setUpAwsConfig();
 
 // Imports
 import { NextApiRequest, NextApiResponse } from "next";
 import httpMocks from "node-mocks-http";
-import {
-  S3Event,
-  S3Object,
-  SNSMessage,
-} from "../app/apis/catalog/hca-atlas-tracker/aws/schemas";
 import { METHOD } from "../app/common/entities";
 import { resetConfigCache } from "../app/config/aws-resources";
 import { endPgPool, query } from "../app/services/database";
@@ -37,114 +44,6 @@ const { httpGet } = jest.requireMock("../app/utils/http") as {
   httpGet: jest.Mock;
 };
 const mockHttpGet = httpGet as jest.Mock;
-
-// Test file path constants
-const TEST_PATH_SEGMENTS = {
-  GUT_V1_INTEGRATED_OBJECTS: "gut/gut-v1/integrated-objects",
-  GUT_V1_MANIFESTS: "gut/gut-v1/manifests",
-  GUT_V1_SOURCE_DATASETS: "gut/gut-v1/source-datasets",
-} as const;
-
-const TEST_FILE_PATHS = {
-  INTEGRATED_OBJECT: `${TEST_PATH_SEGMENTS.GUT_V1_INTEGRATED_OBJECTS}/atlas.h5ad`,
-  MANIFEST: `${TEST_PATH_SEGMENTS.GUT_V1_MANIFESTS}/upload-manifest.json`,
-  SOURCE_DATASET_AUTH: `${TEST_PATH_SEGMENTS.GUT_V1_SOURCE_DATASETS}/auth-test.h5ad`,
-  SOURCE_DATASET_DUPLICATE: `${TEST_PATH_SEGMENTS.GUT_V1_SOURCE_DATASETS}/duplicate-test.h5ad`,
-  SOURCE_DATASET_ETAG: `${TEST_PATH_SEGMENTS.GUT_V1_SOURCE_DATASETS}/etag-test.h5ad`,
-  SOURCE_DATASET_TEST: `${TEST_PATH_SEGMENTS.GUT_V1_SOURCE_DATASETS}/test-file.h5ad`,
-  SOURCE_DATASET_VERSIONED: `${TEST_PATH_SEGMENTS.GUT_V1_SOURCE_DATASETS}/versioned-file.h5ad`,
-} as const;
-
-// SQL query constants
-const SQL_QUERIES = {
-  SELECT_FILE_BY_BUCKET_AND_KEY:
-    "SELECT * FROM hat.files WHERE bucket = $1 AND key = $2",
-  SELECT_FILE_BY_BUCKET_AND_KEY_ORDERED:
-    "SELECT * FROM hat.files WHERE bucket = $1 AND key = $2 ORDER BY created_at ASC",
-  SELECT_LATEST_FILE_BY_BUCKET_AND_KEY:
-    "SELECT * FROM hat.files WHERE bucket = $1 AND key = $2 AND is_latest = true",
-} as const;
-
-// SNS message constants
-const SNS_MESSAGE_DEFAULTS = {
-  SUBJECT: "Amazon S3 Notification",
-} as const;
-
-// Test data constants
-const TEST_VERSION_IDS = {
-  DEFAULT: "096fKKXTRTtl3on89fVO.nfljtsv6qko",
-} as const;
-
-const TEST_TIMESTAMP = "2024-01-01T12:00:00.000Z";
-const TEST_TIMESTAMP_ALT = "2023-01-01T00:00:00.000Z";
-const TEST_SIGNATURE = "fake-signature-for-testing";
-const TEST_SIGNATURE_VALID = "valid-signature";
-const TEST_SIGNATURE_INVALID = "INVALID-SIGNATURE-SHOULD-BE-REJECTED";
-const TEST_ERROR_MESSAGE_INVALID_SIGNATURE = "Invalid signature";
-const TEST_MODULE_SNS_VALIDATOR = "sns-validator";
-
-// S3 Event and SNS Message Factory Functions
-interface S3EventOptions {
-  bucket?: string;
-  etag: string;
-  eventName?: string;
-  eventTime?: string;
-  key: string;
-  sha256?: string; // Optional to support tests that intentionally omit SHA256
-  size: number;
-  versionId: string;
-}
-
-function createS3Event(options: S3EventOptions): S3Event {
-  const objectData: S3Object = {
-    eTag: options.etag,
-    key: options.key,
-    size: options.size,
-    versionId: options.versionId,
-  };
-
-  return {
-    Records: [
-      {
-        eventName: options.eventName || TEST_S3_EVENT_NAME,
-        eventSource: "aws:s3",
-        eventTime: options.eventTime || TEST_TIMESTAMP,
-        s3: {
-          bucket: {
-            name: options.bucket || TEST_S3_BUCKET,
-          },
-          object: objectData,
-        },
-      },
-    ],
-  };
-}
-
-interface SNSMessageOptions {
-  messageId?: string;
-  s3Event: S3Event;
-  signature?: string;
-  signingCertURL?: string;
-  subject?: string;
-  timestamp?: string;
-  topicArn?: string;
-}
-
-function createSNSMessage(options: SNSMessageOptions): SNSMessage {
-  return {
-    Message: JSON.stringify(options.s3Event),
-    MessageId: options.messageId || "test-message-id",
-    Signature: options.signature || TEST_SIGNATURE,
-    SignatureVersion: "1",
-    SigningCertURL:
-      options.signingCertURL ||
-      "https://sns.us-east-1.amazonaws.com/SimpleNotificationService-fake.pem",
-    Subject: options.subject || SNS_MESSAGE_DEFAULTS.SUBJECT,
-    Timestamp: options.timestamp || TEST_TIMESTAMP,
-    TopicArn: options.topicArn || TEST_AWS_CONFIG.sns_topics[0],
-    Type: "Notification",
-  };
-}
 
 jest.mock(
   "../site-config/hca-atlas-tracker/local/authentication/next-auth-config"
@@ -164,74 +63,13 @@ afterEach(() => {
 
 jest.mock("sns-validator", () => {
   return jest.fn().mockImplementation(() => ({
-    validate: jest.fn((message, callback) => {
-      // Check if this is our test case for invalid signatures
-      if (message.Signature === TEST_SIGNATURE_INVALID) {
-        callback(new Error(TEST_ERROR_MESSAGE_INVALID_SIGNATURE));
-        return;
-      }
-
-      // For all other test cases, simulate successful validation
-      callback(null, message);
-    }),
+    validate: jest.fn(validateTestSnsMessage),
   }));
 });
 
 const TEST_ROUTE = "/api/sns";
 
 import snsHandler from "../pages/api/sns";
-
-// Helper function to create test atlas data
-async function createTestAtlasData(): Promise<void> {
-  // Create multiple test atlases to cover different network/version scenarios
-  const atlases = [
-    {
-      id: TEST_GUT_ATLAS_ID,
-      overview: {
-        description: "Test gut atlas for S3 notification integration tests",
-        network: "gut", // Matches S3 path 'gut/gut-v1/...'
-        shortName: "Gut", // Case-insensitive match with S3 atlas base name 'gut'
-        title: "Test Gut Atlas v1",
-        version: "1", // DB version format (S3 'v1' -> DB '1')
-      },
-    },
-    {
-      id: "550e8400-e29b-41d4-a716-446655440001",
-      overview: {
-        description: "Test retina atlas for S3 notification integration tests",
-        network: "eye", // Matches S3 path 'eye/retina-v1/...'
-        shortName: "Retina", // Case-insensitive match with S3 atlas base name 'retina'
-        title: "Test Retina Atlas v1",
-        version: "1", // DB version format (S3 'v1' -> DB '1')
-      },
-    },
-    {
-      id: "550e8400-e29b-41d4-a716-446655440002",
-      overview: {
-        description:
-          "Test gut atlas v1.1 for S3 notification integration tests",
-        network: "gut", // Same network as first gut atlas
-        shortName: "Gut", // Same shortName but different version
-        title: "Test Gut Atlas v1.1",
-        version: "1.1", // DB version format (S3 'v1-1' -> DB '1.1')
-      },
-    },
-  ];
-
-  // Insert all test atlases
-  for (const atlas of atlases) {
-    await query(
-      `INSERT INTO hat.atlases (id, overview, source_studies, status, created_at, updated_at)
-       VALUES ($1, $2, $3, $4, NOW(), NOW())`,
-      [
-        atlas.id,
-        JSON.stringify(atlas.overview),
-        JSON.stringify([]), // Empty source studies array
-        "draft", // Status field
-      ]
-    );
-  }
-}
 
 beforeEach(async () => {
   await resetDatabase();
